delete theme folder
